update readme and rpn_quantize.sh, in order not to influence normal training and inference task.

+++ tools/scripts/rpn_quantize.sh
@@ -13,43 +13,59 @@ source activate torch1.3.1
 pwd
 pwddir=${PWD} #tools
 parentdir="$(dirname "${pwddir}")" #pcdet
-eval_or_calib=`cat $parentdir/pcdet/pointpillar_quantize_config/rpn_quantize_config.json | \
-    python -c 'import json,sys;obj=json.load(sys.stdin);print(obj["eval_or_calib"])'`;
 
-echo $eval_or_calib
-if [ "$eval_or_calib" = "eval" ]; then
-  python test_evaluate_cpp_result.py --cfg_file cfgs/neolix_models/pointpillar_1022_for_calib.yaml --ckpt /home/songhongli/1274_pcdet/checkpoint_pth/checkpoint_epoch_80_1022.pth --batch_size 1 --workers 1
-else
-  #step2. 生成量化表
-  # optional: 选择校准数据集，如果之后的数据集没有变化，这一步骤运行时输入除了 ‘y’之外的任意键即可
-  python $pwddir/calib_utils/cal_data_distrib.py
-  python $pwddir/calib_utils/mv_calib_datasets.py
+quantize_mode=`cat $parentdir/pcdet/pointpillar_quantize_config/rpn_quantize_config.json | \
+    python -c 'import json,sys;obj=json.load(sys.stdin);print(obj["quantize_mode"])'`;
 
-  copied_calib_pc_dir=`cat $parentdir/pcdet/pointpillar_quantize_config/rpn_quantize_config.json | \
-      python -c 'import json,sys;obj=json.load(sys.stdin);print(obj["copied_calib_pc_dir"])'`;
+if [ "$quantize_mode" = "true" ]; then
+  eval_or_calib=`cat $parentdir/pcdet/pointpillar_quantize_config/rpn_quantize_config.json | \
+      python -c 'import json,sys;obj=json.load(sys.stdin);print(obj["eval_or_calib"])'`;
 
-  # --------- 清理掉之前的文件 ---------
-  calib_rpn_input_dir=`cat $parentdir/pcdet/pointpillar_quantize_config/rpn_quantize_config.json | \
-      python -c 'import json,sys;obj=json.load(sys.stdin);print(obj["calib_rpn_input_dir"])'`;
-  echo $calib_rpn_input_dir
-  rm -rf $calib_rpn_input_dir/*.*
-  # --------^ 清理掉之前的文件 ^--------
+  dataset_yaml_file=`cat $parentdir/pcdet/pointpillar_quantize_config/rpn_quantize_config.json | \
+      python -c 'import json,sys;obj=json.load(sys.stdin);print(obj["dataset_yaml_file"])'`;
 
-  python demo_for_calib.py --cfg_file cfgs/neolix_models/pointpillar_1022_for_calib.yaml --ckpt /home/songhongli/1274_pcdet/checkpoint_pth/checkpoint_epoch_80_1022.pth --data_path $copied_calib_pc_dir --output_path /home/songhongli/1274_pcdet/infer_out_useless/
+  checkpoint_pth=`cat $parentdir/pcdet/pointpillar_quantize_config/rpn_quantize_config.json | \
+      python -c 'import json,sys;obj=json.load(sys.stdin);print(obj["checkpoint_pth"])'`;
 
-  python $pwddir/calib_utils/generate_calib_list.py
+  echo $eval_or_calib
+  if [ "$eval_or_calib" = "eval" ]; then
+      python test_evaluate_cpp_result.py --cfg_file cfgs/neolix_models/$dataset_yaml_file --ckpt $checkpoint_pth --batch_size 1 --workers 1
+  else
+      #step2. 生成量化表
+      # optional: 选择校准数据集，如果之后的数据集没有变化，这一步骤运行时输入除了 ‘y’之外的任意键即可
+      python $pwddir/calib_utils/cal_data_distrib.py
+      python $pwddir/calib_utils/mv_calib_datasets.py
 
-  python $pwddir/onnx_utilis/export_vfe_weight.py
+      copied_calib_pc_dir=`cat $parentdir/pcdet/pointpillar_quantize_config/rpn_quantize_config.json | \
+          python -c 'import json,sys;obj=json.load(sys.stdin);print(obj["copied_calib_pc_dir"])'`;
+      # --------- 清理掉之前的文件 ---------
+      calib_rpn_input_dir=`cat $parentdir/pcdet/pointpillar_quantize_config/rpn_quantize_config.json | \
+          python -c 'import json,sys;obj=json.load(sys.stdin);print(obj["calib_rpn_input_dir"])'`;
+      echo $calib_rpn_input_dir
+      rm -rf $calib_rpn_input_dir/*.*
+      # --------^ 清理掉之前的文件 ^--------
 
-  vfe_exported_weight_file=`cat $parentdir/pcdet/pointpillar_quantize_config/rpn_quantize_config.json | \
-      python -c 'import json,sys;obj=json.load(sys.stdin);print(obj["vfe_exported_weight_file"])'`;
+      python demo_for_calib.py --cfg_file cfgs/neolix_models/$dataset_yaml_file --ckpt $checkpoint_pth --data_path $copied_calib_pc_dir
 
-  rpn_tarfile_name=`cat $parentdir/pcdet/pointpillar_quantize_config/rpn_quantize_config.json | \
-      python -c 'import json,sys;obj=json.load(sys.stdin);print(obj["rpn_tarfile_name"])'`;
+      python $pwddir/calib_utils/generate_calib_list.py
 
-  echo "ready to scp to xavier: check the file:"
-  echo $rpn_tarfile_name
-  echo $vfe_exported_weight_file
+      python $pwddir/onnx_utilis/export_vfe_weight.py
 
-  echo "done!"
+      vfe_exported_weight_file=`cat $parentdir/pcdet/pointpillar_quantize_config/rpn_quantize_config.json | \
+          python -c 'import json,sys;obj=json.load(sys.stdin);print(obj["vfe_exported_weight_file"])'`;
+
+      rpn_tarfile_name=`cat $parentdir/pcdet/pointpillar_quantize_config/rpn_quantize_config.json | \
+          python -c 'import json,sys;obj=json.load(sys.stdin);print(obj["rpn_tarfile_name"])'`;
+
+      echo "ready to scp to xavier: check the file:"
+      echo $rpn_tarfile_name
+      echo $vfe_exported_weight_file
+
+      echo "done!"
+  fi
+else  
+    echo "according to rpn_quantize_config.json, the quantize_mode was set as false, so we close this script."
 fi
+
+echo "script done!"
+
